requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
opencv-python
scikit-learn
matplotlib
streamlit
requests